requirements.txt
pyodbc>=4.0.39
mysql-connector-python>=8.0.32
pytest>=7.4
pytest-mock>=3.11
